requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==7.1.2
colorama==0.4.4
dataclasses==0.8
Flask==1.1.4
flask-restx==0.4.0
importlib-metadata==4.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
PyJWT==2.1.0
pyrsistent==0.17.3
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==1.0.1
zipp==3.4.1
